+++ app/javascript/lca/components/battlegroups/BattlegroupCard.jsx
@@ -11,18 +11,19 @@ import DragHandleIcon from '@material-ui/icons/DragHandle'
 import Launch from '@material-ui/icons/Launch'
 import VisibilityOff from '@material-ui/icons/VisibilityOff'
 
+import BattlegroupHealthDisplay from './BattlegroupHealthDisplay.jsx'
 import PlayerNameSubtitle from '../generic/PlayerNameSubtitle.jsx'
-import PoolDisplay from '../generic/PoolDisplay.jsx'
 import CharacterCardMenu from '../generic/CharacterCardMenu'
-import ResourceDisplay from '../generic/ResourceDisplay.jsx'
+import sharedStyles from 'styles/'
 import { doIOwnBattlegroup } from 'selectors'
-import { prettyDrillRating, totalMagnitude } from 'utils/calculated'
+import { prettyDrillRating } from 'utils/calculated'
 
 const Handle = SortableHandle(() => (
   <DragHandleIcon onClick={e => e.preventDefault()} />
 ))
 
 const styles = theme => ({
+  ...sharedStyles(theme),
   root: {
     ...theme.mixins.gutters({
       paddingTop: 16,
@@ -37,9 +38,6 @@ const styles = theme => ({
     verticalAlign: 'middle',
     lineHeight: 'inherit',
   },
-  nameRow: {
-    display: 'flex',
-  },
   nameWrap: {
     flex: 1,
   },
@@ -49,10 +47,6 @@ const styles = theme => ({
   icon: {
     verticalAlign: 'bottom',
     marginLeft: theme.spacing.unit,
-  },
-  rowContainer: {
-    display: 'flex',
-    flexWrap: 'wrap',
   },
   statWrap: {
     marginRight: theme.spacing.unit,
@@ -92,7 +86,7 @@ function BattlegroupCard(props: Props) {
         </Typography>
       )}
 
-      <div className={classes.nameRow}>
+      <div className={classes.flexContainer}>
         <div className={classes.nameWrap}>
           <Typography
             variant="title"
@@ -119,24 +113,14 @@ function BattlegroupCard(props: Props) {
         )}
       </div>
 
-      <div className={classes.rowContainer}>
-        <ResourceDisplay
-          current={battlegroup.magnitude}
-          total={totalMagnitude(battlegroup)}
-          label="Magnitude"
+      <div className={classes.flexContainerWrap}>
+        <BattlegroupHealthDisplay
+          battlegroup={battlegroup}
           className={classes.statWrap}
         />
 
-        <PoolDisplay
-          battlegroup
-          staticRating
-          pool={{ total: battlegroup.size }}
-          label="Size"
-          classes={{ root: classes.statWrap }}
-        />
-
         <div className={classes.statWrap}>
-          <div className={classes.statLabel}>Drill:</div>
+          <div className={classes.statLabel}>Drill</div>
           <div className={classes.statValue}>
             {prettyDrillRating(battlegroup)}
           </div>
@@ -144,13 +128,13 @@ function BattlegroupCard(props: Props) {
 
         {battlegroup.might > 0 && (
           <div className={classes.statWrap}>
-            <div className={classes.statLabel}>Might:</div>
+            <div className={classes.statLabel}>Might</div>
             <div className={classes.statValue}>{battlegroup.might}</div>
           </div>
         )}
         {battlegroup.perfect_morale && (
           <div className={classes.statWrap}>
-            <div className={classes.statLabel}>Morale:</div>
+            <div className={classes.statLabel}>Morale</div>
             <div className={classes.statValue}>Perfect</div>
           </div>
         )}